+++ example/lib/utils.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'main.dart';
 
 Future asyncInputDialog(BuildContext context, String prompt,