+++ geofence/android/src/main/kotlin/com/moengage/flutter/geofence/Constants.kt
@@ -1,0 +1,6 @@
+package com.moengage.flutter.geofence
+
+internal const val CHANNEL_NAME = "com.moengage/geofence"
+
+internal const val METHOD_NAME_START_GEOFENCE_MONITORING = "startGeofenceMonitoring"
+internal const val METHOD_NAME_STOP_GEOFENCE_MONITORING = "stopGeofenceMonitoring"