+++ geofence/lib/moe_android_geofence.dart
@@ -1,22 +1,21 @@
 import 'package:flutter/services.dart';
-import 'package:moengage_flutter/internal/logger.dart';
+import 'package:moengage_flutter/moengage_flutter.dart';
 
 import 'moe_geofence_constants.dart';
 import 'moe_geofence_utils.dart';
 
-class MoEiOSGeofence {
-  String _tag = "${TAG}MoEiOSGeofence";
+class MoEAndroidGeofence {
+  String _tag = "${TAG}MoEAndroidGeofence";
   MethodChannel _channel;
 
-  MoEiOSGeofence(this._channel);
+  MoEAndroidGeofence(this._channel);
 
   void startGeofenceMonitoring(String appId) {
     try {
       _channel.invokeMethod(
           methodStartGeofenceMonitoring, getAccountMeta(appId));
-    } catch (e, stacktrace) {
-      Logger.e("$_tag Error: startGeofenceMonitoring() : ",
-          error: e, stackTrace: stacktrace);
+    } catch (e) {
+      Logger.e("$_tag Error: startGeofenceMonitoring() : $e");
     }
   }